backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
aiofiles==23.2.1
